requirements.txt
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
psycopg2
python-twitter
oauth2